Add 8-bit decoder and use it for tests
This is LUT-based decoder, decoding 4 bits at a time. The LUT is tiny. I
will try a similar approach with a larger LUT and 8 bits at a time; LUT
size will be much larger, but still reasonable (2KiB?). It will have to
contain 64-bit integers, which will be most efficient on 64-bit
machines, but workable on 32-bit machines as well.

The inputs are pre-interleaved in `mk-inputs.py`. Input interleaver in C
is to be written soon. The test rig needs to be rearranged to easily
support testing different setups: no interleave, 8-bit interleave, etc.