Fix 64-bit version
The bug was in packing. Gnarly. Turns out it was interpreting 0x1 as
a signed 32-bit constant and thus emitting a 32-bit shl and
sign-extending it to 64 bits, which broke the upper 32 bits of my data
in horrible ways.

Fixed by passing a properly typed constant for the 64-bit version.